// src/app/modules/report-input/models/ReportModels.ts
export interface Inpatient {
  id: number;
  soBHYT: string;
  soCMNDOrCCCD: string;
  tuyenKCB: string;
  gender: string;
  age: number;
  ageToHospital: number;
  birthDate: string;
  object: string;
  treatmentRoom: string;
  treatmentChamber: string;
  treatmentBed: string;
  PhysicianName: string;
  maBenh: string;
  tenBenh: string;
  maBenhKemTheo: string;
  tenBenhKemTheo: string;
  diaChi: string;
}
export interface formData {
  fromDate: string;
  toDate: string;
  khoaBaoCao: any | null;
  maKhoaPhong: string;
  dieuDuongTongSo: number | null;
  dieudDuongHienCo: number | null;
  dieuDuongVang: number | null;
  dieuDuongTrucTongSo: number | null;
  dieuDuongTruc: string[] | null;
  dieuDuongNghiOmTongSo: number | null;
  dieuDuongNghiOm: string[] | null;
  dieuDuongNghiTrucTongSo: number | null;
  dieuDuongNghiTruc: string[] | null;
  dieuDuongCongTacTongSo: number | null;
  dieuDuongCongTac: string[] | null;
  dieuDuongDiHocTongSo: number | null;
  danhSachDiHoc: string[] | null;
  dieuDuongChoHuuTongSo: number | null;
  dieuDuongChoHuu: string[] | null;
  dieuDuongNghiKhacTongSo: number | null;
  dieuDuongNghiKhac: string[] | null;
  soLuongDangCongTac: number | null;
  soLuongNghiHuu: number | null;
}

export interface BenhNhan {
  createdDate: string | null;
  createdBy: string | null;
  modifiedDate: string | null;
  modifiedBy: string | null;
  id: string | null;
  maBenhNhan: string | null;
  maVienPhi: string | null;
  hoTen: string | null;
  donTiepCanCuocCongDan: string | null;
  soBaoHiemYTe: string | null;
  canCuocCongDan: string | null;
  tuyenKhamChuaBenh: number | null;
  gioiTinh: number | null;
  tuoi: number | null;
  tuoiVaoVien: number | null;
  ngaySinh: string | null;
  doiTuong: number | null;
  doiTuongBaoHiemYTe: number | null;
  maBenh: string | null;
  tenBenh: string | null;
  maBenhKemTheo: string | null;
  tenBenhKemTheo: string | null;
  diaChi: string | null;
  hoTenBo: string | null;
  hoTenMe: string | null;
  hoTenNguoiNha: string | null;
  soDienThoaiBo: string | null;
  soDienThoaiMe: string | null;
  soDienThoaiNguoiNha: string | null;
  vaoTuKhoa: string | null;
  ngayVaoKhoa: string | null;
  ngayRaKhoa: string | null;
  hinhThucVaoKhoa: number | null;
  phongDieuTri: string | null;
  buongDieuTri: string | null;
  giuongDieuTri: string | null;
  bacsiDieuTri: string | null;
  ketQuaDieuTri: string | null;
  soNgayDieuTri: number | null;
  xuTri: number | null;
  chuyenDenKhoa: number | null;
  benhVienChuyenDi: string | null;
  ngayVaoVien: string | null;
  ngayDieuTriTiep: string | null;
}